+++ Com.DanLiris.Service.Core.Lib/ViewModels/PackingDetailModel.cs
@@ -12,8 +12,8 @@ namespace Com.DanLiris.Service.Core.Lib.ViewModels
         public string Lot { get; set; }
         [MaxLength(100)]
         public string Grade { get; set; }
-        public int Weight { get; set; }
-        public int Length { get; set; }
+        public double Weight { get; set; }
+        public double Length { get; set; }
         public int Quantity { get; set; }
         [MaxLength(500)]
         public string Remark { get; set; }
